등록버튼 권한설정 1차 완료

+++ dip_mes/Standard/Product.cs
@@ -145,6 +145,11 @@ namespace dip_mes
 
         private void btnRegister_Click(object sender, EventArgs e)
         {
+            if (Login.getAuth != 1 || Login.getAuth != 3)
+            {
+                MessageBox.Show("권한이 없습니다.");
+                return;
+            }
             // 데이터 그리드뷰에 데이터 추가
             DateTime ins_date = DateTime.Now;
 
@@ -199,6 +204,11 @@ namespace dip_mes
 
         private void InsertProductProcess(string processName, int processTime, string productCode)
         {
+            if (Login.getAuth != 1 || Login.getAuth != 3)
+            {
+                MessageBox.Show("권한이 없습니다.");
+                return;
+            }
             // product_process 테이블에 사용자에게 입력받은 데이터 DB 저장
             string Query = "INSERT INTO product_process (product_code, process_name, process_time) VALUES (@product_code, @process_name, @process_time)";
             using (MySqlCommand Cmd = new MySqlCommand(Query, connection))
@@ -220,6 +230,11 @@ namespace dip_mes
         }
         private void InsertProductMaterial(string MaterialName, int MaterialNumber, string productCode)
         {
+            if (Login.getAuth != 1 || Login.getAuth != 3)
+            {
+                MessageBox.Show("권한이 없습니다.");
+                return;
+            }
             // product_Material 테이블에 사용자에게 입력받은 데이터 DB 저장
             string Query = "INSERT INTO product_Material (product_code, Material_name, Material_number) VALUES (@product_code, @Material_name, @Material_number)";
             using (MySqlCommand Cmd = new MySqlCommand(Query, connection))
@@ -256,6 +271,11 @@ namespace dip_mes
 
         private void InsertDataIntoMySQL(string product_code, string product_name, string category, string product_size, DateTime ins_date)
         {
+            if (Login.getAuth != 1 || Login.getAuth != 3)
+            {
+                MessageBox.Show("권한이 없습니다.");
+                return;
+            }
             string query = "INSERT INTO product (product_code, product_name, category, product_size, ins_date) VALUES (@product_code, @product_name, @category, @product_size, @ins_date)";
             using (MySqlCommand cmd = new MySqlCommand(query, connection))
             {
@@ -296,12 +316,18 @@ namespace dip_mes
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
+            if (Login.getAuth != 1 || Login.getAuth != 3)
+            {
+                MessageBox.Show("권한이 없습니다.");
+                return;
+            }
             // 체크된 항목 삭제
             DeleteCheckedItems();
         }
 
         private void DeleteCheckedItems()
         {
+
             // 체크된 항목을 반복하여 삭제
             foreach (DataGridViewRow row in dataGridView1.Rows)
             {
@@ -511,6 +537,11 @@ namespace dip_mes
 
         private void btnRegister1_Click(object sender, EventArgs e)
         {
+            if (Login.getAuth != 1 || Login.getAuth != 3)
+            {
+                MessageBox.Show("권한이 없습니다.");
+                return;
+            }
             if (string.IsNullOrEmpty(textbox6.Text)) 
             {
                 MessageBox.Show("먼저 제품정보를 조회후 선택하세요");
@@ -533,6 +564,11 @@ namespace dip_mes
 
         private void button3_Click(object sender, EventArgs e)
         {
+            if (Login.getAuth != 1 || Login.getAuth != 3)
+            {
+                MessageBox.Show("권한이 없습니다.");
+                return;
+            }
             if (string.IsNullOrEmpty(textBox7.Text))
             {
                 MessageBox.Show("먼저 제품정보를 조회후 선택하세요");
@@ -555,6 +591,11 @@ namespace dip_mes
 
         private void btnDelete1_Click(object sender, EventArgs e)
         {
+            if (Login.getAuth != 1 || Login.getAuth != 3)
+            {
+                MessageBox.Show("권한이 없습니다.");
+                return;
+            }
             // 체크된 항목을 반복하여 삭제
             foreach (DataGridViewRow row in dataGridView2.Rows)
             {
